+++ busy_beaver/apps/slack_integration/blocks.py
@@ -1,4 +1,4 @@
-from busy_beaver.apps.events.workflow import generate_upcoming_events_message
+from busy_beaver.apps.events.upcoming_events import generate_upcoming_events_message
 from busy_beaver.toolbox.slack_block_kit import Divider, Section
 from busy_beaver.toolbox.slack_block_kit.blocks import Block
